Fix for y2k38 issue and integer overflow

+++ src/rtmessage/rtrouteBase.h
@@ -31,9 +31,9 @@ typedef struct
   uint8_t*                  read_buffer;
   uint8_t*                  send_buffer;
   rtConnectionState         state;
-  int                       bytes_read;
-  int                       bytes_to_read;
-  int                       read_buffer_capacity;
+  uint64_t                       bytes_read;
+  uint64_t                       bytes_to_read;
+  uint64_t                       read_buffer_capacity;
   rtMessageHeader           header;
 #ifdef WITH_SPAKE2
   rtCipher*                 cipher;
